Refactor Viewposts
Refactored viewposts into seperate file
